refactor(webfont-generator): establish format encoders
