+++ packages/webfont-generator/src/formats/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod eot;
+pub(crate) mod woff1;
